services/analysis/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-Levenshtein==0.25.0
confusable_homoglyphs==3.2.0
semgrep==1.66.0
dnspython==2.6.1
requests==2.31.0
